deprecate the ProviderFactoryTestCase
