+++ src/Symfony/Component/Translation/Bridge/Loco/composer.json
@@ -19,7 +19,7 @@
         "php": ">=8.2",
         "symfony/http-client": "^6.4|^7.0",
         "symfony/config": "^6.4|^7.0",
-        "symfony/translation": "^6.4|^7.0"
+        "symfony/translation": "^7.2"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Translation\\Bridge\\Loco\\": "" },
