+++ src/Symfony/Component/Translation/CHANGELOG.md
@@ -4,6 +4,11 @@ CHANGELOG
 7.2
 ---
 
+ * Deprecate `ProviderFactoryTestCase`, extend `AbstractTransportFactoryTestCase` instead
+
+   The `testIncompleteDsnException()` test is no longer provided by default. If you make use of it by implementing the `incompleteDsnProvider()` data providers,
+   you now need to use the `IncompleteDsnTestTrait`.
+
  * Make `ProviderFactoryTestCase` and `ProviderTestCase` compatible with PHPUnit 10+
  * Add `lint:translations` command
  * Deprecate passing an escape character to `CsvFileLoader::setCsvControl()`
