+++ src/Symfony/Component/Translation/Test/ProviderFactoryTestCase.php
@@ -11,17 +11,11 @@
 
 namespace Symfony\Component\Translation\Test;
 
-use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\MockObject\MockObject;
-use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpClient\MockHttpClient;
 use Symfony\Component\Translation\Dumper\XliffFileDumper;
-use Symfony\Component\Translation\Exception\IncompleteDsnException;
-use Symfony\Component\Translation\Exception\UnsupportedSchemeException;
 use Symfony\Component\Translation\Loader\LoaderInterface;
-use Symfony\Component\Translation\Provider\Dsn;
-use Symfony\Component\Translation\Provider\ProviderFactoryInterface;
 use Symfony\Component\Translation\TranslatorBagInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
@@ -29,27 +23,19 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
  * A test case to ease testing a translation provider factory.
  *
  * @author Mathieu Santostefano <msantostefano@protonmail.com>
+ *
+ * @deprecated since Symfony 7.2, use AbstractProviderFactoryTestCase instead
  */
-abstract class ProviderFactoryTestCase extends TestCase
+abstract class ProviderFactoryTestCase extends AbstractProviderFactoryTestCase
 {
+    use IncompleteDsnTestTrait;
+
     protected HttpClientInterface $client;
     protected LoggerInterface|MockObject $logger;
     protected string $defaultLocale;
     protected LoaderInterface|MockObject $loader;
     protected XliffFileDumper|MockObject $xliffFileDumper;
     protected TranslatorBagInterface|MockObject $translatorBag;
-
-    abstract public function createFactory(): ProviderFactoryInterface;
-
-    /**
-     * @return iterable<array{0: bool, 1: string}>
-     */
-    abstract public static function supportsProvider(): iterable;
-
-    /**
-     * @return iterable<array{0: string, 1: string}>
-     */
-    abstract public static function createProvider(): iterable;
 
     /**
      * @return iterable<array{0: string, 1?: string|null}>
@@ -65,65 +51,6 @@ abstract class ProviderFactoryTestCase extends TestCase
     public static function incompleteDsnProvider(): iterable
     {
         return [];
-    }
-
-    /**
-     * @dataProvider supportsProvider
-     */
-    #[DataProvider('supportsProvider')]
-    public function testSupports(bool $expected, string $dsn)
-    {
-        $factory = $this->createFactory();
-
-        $this->assertSame($expected, $factory->supports(new Dsn($dsn)));
-    }
-
-    /**
-     * @dataProvider createProvider
-     */
-    #[DataProvider('createProvider')]
-    public function testCreate(string $expected, string $dsn)
-    {
-        $factory = $this->createFactory();
-        $provider = $factory->create(new Dsn($dsn));
-
-        $this->assertSame($expected, (string) $provider);
-    }
-
-    /**
-     * @dataProvider unsupportedSchemeProvider
-     */
-    #[DataProvider('unsupportedSchemeProvider')]
-    public function testUnsupportedSchemeException(string $dsn, ?string $message = null)
-    {
-        $factory = $this->createFactory();
-
-        $dsn = new Dsn($dsn);
-
-        $this->expectException(UnsupportedSchemeException::class);
-        if (null !== $message) {
-            $this->expectExceptionMessage($message);
-        }
-
-        $factory->create($dsn);
-    }
-
-    /**
-     * @dataProvider incompleteDsnProvider
-     */
-    #[DataProvider('incompleteDsnProvider')]
-    public function testIncompleteDsnException(string $dsn, ?string $message = null)
-    {
-        $factory = $this->createFactory();
-
-        $dsn = new Dsn($dsn);
-
-        $this->expectException(IncompleteDsnException::class);
-        if (null !== $message) {
-            $this->expectExceptionMessage($message);
-        }
-
-        $factory->create($dsn);
     }
 
     protected function getClient(): HttpClientInterface
